+++ CHANGES.md
@@ -5,7 +5,9 @@
 * adding vacation information to JSON output of summary subcommand
 * adding precision and truncation to summary subcommand
 * bumped colonnade version to get better grapheme handling
-* adding hours clocked to stastics and allowing one to collect statistics by time period
+* adding hours clocked to statistics and allowing one to collect statistics by time period
+* made 12 vs 24 hour clock configurable and removed am/pm
+* fixed time display so colons line up
 ## 0.1.4 *2020-1-26*
 * fix use of max_termsize so we can test things outside of tty
 * added 'today' alias to the summary subcommand for Paula's sake